backend/requirements.txt
openai
flask
flask-cors
flask-sqlalchemy
python-dotenv
pydantic-settings
aspose-words
python-docx